refactor: split App.js into modular files
