+++ frontend/src/constants.js
@@ -1,0 +1,30 @@
+export const PRODUCTS = [
+  'Enviroshake',
+  'Enviroslate',
+  'Enviroshingle'
+];
+
+export const WAREHOUSES = ['Dresden', 'BC', 'Buffalo'];
+
+export const TYPES = [
+  'Bundle',
+  'Cap 2-3',
+  'Cap 4-5',
+  'Cap 6-7',
+  'Cap 8-9',
+  'Cap 10-11',
+  'Cap 12-13',
+  'Cap 14-15',
+  'Cap 16-17',
+  'Cap 18-19',
+  'Cap 20-21'
+];
+
+export const STAGES = [
+  'Available',
+  'Open',
+  'Released',
+  'Staged',
+  'Shipped',
+  'Transfer'
+];
